Remove 'get' prefix from bika catalog indexes!
This reduces the confusion caused by not being able to distinguish
an accessor from an index or a catalog metadata column.

Now at a moments glance, in any context, we know that:

xxx.getSomething() or xxx/getSomething: definitely an object/accessor.
xxx.Something or xxx/Something: definitely an Index or proxy metadata.

Apologies for the upheaval this will surely cause.